feat: make env loop wrappers behave like wrappers
Environment loop wrappers are currently sub-classes
of ParallelEnvironmentLoop. This prevents the use of
custom environment loops since all systems apply a
statistics wrapper(inherits from ParallelEnnvironmentLoop)
to any suppled environment loop.

This behaviour is solved by making EnvironmentLoopStatisticsBase
not to inherit from ParallelEnvironmentLoop,
but making it to wrap the supplied environment
loop and override the environment loop's
methods. Also an attribute getter is introduced
which looks for attribute in the environment loop
whenever it is missing in the wrapper.