Add return metrics for experiments and corresponding Hydra config option
Modifies the run_experiment method and hydra_entry_point in the various systems to return a metric to be used for hyperparameter tuning through Hydra (using something like OpTuna). The Hydra config option, config.env.eval_metric,is used to select the metric (episode_return or win_rate) for an environment.